Removed bad namespace and moved command and argument classes up a directory.
